change new resize method arguments to uint32_t type

+++ sensor_msgs/include/sensor_msgs/impl/point_cloud2_iterator.hpp
@@ -144,13 +144,13 @@ inline void PointCloud2Modifier::resize(size_t size)
   }
 }
 
-inline void PointCloud2Modifier::resize(size_t width, size_t height)
+inline void PointCloud2Modifier::resize(uint32_t width, uint32_t height)
 {
   size_t size = width * height;
   resize(size);
 
-  cloud_msg_.width = static_cast<uint32_t>(width);
-  cloud_msg_.height = static_cast<uint32_t>(height);
+  cloud_msg_.width = width;
+  cloud_msg_.height = height;
 }
 
 inline void PointCloud2Modifier::clear()
